This commit should support IaC development
I want to be able to test code changes to the infrastructure. This could be a configuration
change in Kubernetes, IAM role or bucket change, even a newer version of any of the deployed
containers - vault, vault-init, external-dns or cert-manager. I want to see how terraform
will handle the changes, how the infrastructure handles successful and unsuccessful applies
as well as how the applications handle it. I want to run automated local tests, I want to be
somewhat confident when I commit that I will not break the the next environment in the pipeline.